irrelevance woes
can't prove the extensionality that I want because it requires pattern
matching the irrelevant proofs...

I don't get why you can't pattern match refl, the whole point of the
equality type _is_ that there are no other constructors

anyway the other thing is I can recompute surj and inj, maybe I should
make a function for that

but it's kind of tempting to just throw the whole thing out since the
standard library Group and Fin.Permutation have much nicer design in so
many ways

I could still use the functional class, since it is interesting, esp all
the generalizations like functor -> functional
but permutation would just be a group instance for Fin.Permutation, and
then I'd do a subgroup theorem and cayley's theorem?

im planning all of this but i might not do any, it was primarily a
learning project and I have learned heaps.